Suppress the login page
/admin/users/login shows a log in form for users in Omeka 2.5.1.

We don't want local passwords to interfere with cosign logins.
Instead, redirect to /login?dest=/admin instead. We don't need
to preserve the REQUEST_URI because we know it's the log in form.

Since /admin/users/login is in the admin_whitelist, we have to add
our new redirect to the admin_whitelist.

Additionally, the redirect should be sensitive to whether the
browser is using http or https.