Export exhibit metadata in json for solr
We used to use some Drupal 7 views with external tables to get
exhibit metadata into some XML for solr to consume.  But we
retired the vestigial Drupal 7 code in December 2024.  Coincidentally,
Online Exhibits content has been on hold for at least as long.

However, to be prepared for the possibility of exhibits being updated,
I'm making a new mechanism to export exhibit metadata.  That's this.
The metadata is fairly simple: title, url, description, modified,
and tags.